backend/packages/document-ingestion/pyproject.toml
[project]
name = "document-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Fledge Shiu", email = "xzk0701@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langdetect>=1.0.9",
    "playwright>=1.56.0",
    "returns>=0.26.0",
    "trafilatura>=2.0.0",
    "repos",
    "db-models",
    "python-dotenv>=1.2.1",
]

[project.scripts]
document-ingestion = "document_ingestion:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
